api/pyproject.toml
[project]
name = "gomoku-api"
version = "0.1.0"
description = "FastAPI service for Gomoku: auth, game proxy, scoring, leaderboard"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
    "httpx>=0.28",
    "pyjwt>=2.10",
    "bcrypt>=4.2",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.2.2",
    "uuid>=1.30",
    "sqlalchemy>=2.0.49",
    "alembic>=1.18.4",
    "structlog>=25.5.0",
    "logfire>=4.31.0",
    "psycopg2-binary>=2.9.11",
    "opentelemetry-instrumentation-asyncpg>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-instrumentation-urllib>=0.60b1",
    "rich>=14.3.3",
    "typer>=0.24.1",
]
[dependency-groups]
dev = [
    "codespell>=2.4.1",
    "detect-secrets>=1.5.0",
    "gibberish-detector>=0.1.1",
    "icecream>=2.1.8",
    "lefthook>=1.13.0",
    "mdformat-config>=0.2.1",
    "mdformat-gfm>=0.4.1",
    "mdformat-ruff>=0.1.3",
    "mdformat-simple-breaks>=0.0.1",
    "mdformat>=0.7.22",
    "ruff>=0.13.0",
    "ty>=0.0.1a20",
    "uv-sort>=0.6.1",
]
test = [
    "pytest>=9.0.3",
    "pytest-asyncio>=0.24",
    "pytest-coverage>=0.0",
    "httpx",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
